requirements.txt
opencv-python==4.10.0.82
numpy==1.23.4
Pillow==10.3.0